TeleOp Main and other changes
