--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Ports.txt
@@ -1,17 +1,0 @@
-Main Hub
-Port 0 - rightback_drive (par0)
-Port 1 - leftback_drive (par1)
-Port 2 - rightfront_drive (perp)
-Port 3 - leftfront_drive
-1 to 0
-3 to 1
-
-Servo port 0 - elbow_servo
-Servo port 1 - claw_servo
-
-Expansion Hub
-Port 0 - viper_slide_right
-Port 1 - viper_slide_left
-
-All DcMotors are GoBILDA 5202/3/4 series
-
